[Refactor] 글로벌 유틸 함수 위치 변경
