+++ package.json
@@ -11,9 +11,10 @@
   "dependencies": {
     "@nextui-org/accordion": "^2.0.36",
     "bcryptjs": "^2.4.3",
+    "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "dayjs": "^1.11.11",
-    "faker": "^6.6.6",
+    "es-hangul": "^1.5.0",
     "mongoose": "^8.5.0",
     "next": "14.2.4",
     "next-auth": "^5.0.0-beta.19",
@@ -24,10 +25,10 @@
     "react-quill": "^2.0.0",
     "react-select": "^5.8.0",
     "swiper": "^11.1.4",
+    "tailwind-merge": "^2.4.0",
     "uuid": "^10.0.0"
   },
   "devDependencies": {
-    "@faker-js/faker": "^8.4.1",
     "@types/bcryptjs": "^2.4.6",
     "@types/node": "^20",
     "@types/react": "^18",
