+++ src/common/Organisms/PostListItem.tsx
@@ -2,7 +2,7 @@ import Link from "next/link";
 import Keyword from "../Atoms/Text/Keyword";
 import Profile from "../Molecules/Profile";
 import { TPost } from "@/types/model/PostItem";
-import { getCreatedBefore } from "@/dummies/utils";
+import { getCreatedBefore } from "@/utils/getCreatedBefore";
 
 export default function PostListItem({ item }: { item: TPost }) {
   const createdBefore = getCreatedBefore(item.createdAt);
